ZepixTradingBot-old-v2-main/Trading_Bot/requirements.txt
# Zepix Trading Bot v2.0 - Python Dependencies
# Updated: 2026-01-12 (Added Voice Alert System V2.0 dependencies)

# Core Dependencies
python-telegram-bot>=20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
sqlite3  # Built-in with Python

# Phase 2: Fixed Clock System
pytz>=2023.3                    # IST timezone handling

# Phase 5: Voice Alert System V2.0 (Windows Speaker TTS + Telegram Text)
pyttsx3>=2.90                   # Windows offline TTS (NEW - V2.0)

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality (Development)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: SMS Gateway (Production)
# twilio>=8.0.0                 # For SMS fallback alerts
